prepare separate component for header
