+++ packages/boba/gateway/src/containers/Bridging/BridgeHeader/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const BridgeHeaderWrapper = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`
